requirements.txt
beautifulsoup4==4.13.4
Flask==3.1.0
flask_bcrypt==1.0.1
flask_caching==2.3.1
flask_cors==5.0.1
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
groq==0.24.0
langchain_community==0.3.23
ollama==0.4.8
python-dotenv==1.1.0
Requests==2.32.3
soundfile==0.13.1
textblob==0.19.0
